Refactored into a Renderer with ActivePass interfaces

+++ Project1/common.cpp
@@ -1,6 +1,6 @@
 #include "common.h"
 
-byte *read_whole_file(std::string filename, size_t &o_size) {
+byte *read_whole_file(std::string filename, u32 &o_size) {
     std::ifstream file(filename, std::ios::binary);
     
     if (!file.good() || !file.is_open() || file.fail()) {
@@ -10,10 +10,11 @@ byte *read_whole_file(std::string filename, size_t &o_size) {
     }
 
     file.seekg(0, std::fstream::beg);
-    size_t start = file.tellg();
+    u32 start = file.tellg();
     file.seekg(0, std::fstream::end);
-    o_size = (size_t) file.tellg() - start;
+    o_size = (u32) file.tellg() - start;
     file.seekg(0, std::fstream::beg);
+
 
     byte *buf = new byte[o_size];
     memset(buf, 0, o_size);
